% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost_cv.R
\name{boost_CV}
\alias{boost_CV}
\title{Log likelihood based cross validation for iterations in location-scale
boosting models}
\usage{
boost_CV(m, k = 5, mstops, seed = NULL)
}
\arguments{
\item{m}{An object of class \code{lslm_boost} containing the relevant data and
correct settings for the learning rates.}

\item{k}{Number of folds in k-fold cross validation.}

\item{mstops}{Vector of parameters containing positive integer values.}

\item{seed}{Seed for random number generation.}
}
\value{
A list containing numeric values:
\itemize{
\item mean_logliks - The values of log likelihood given the parameter mstop,
averaged over all folds.
\item max_loglik - The value of the biggest average log likelihood.
\item max_mstop - Number of iterations that maximizes the log likelihood.
}
}
\description{
k-fold cross validation for componentwise gradient boosting in location-scale
regression models. The selection criterion is the log likelihood function of
the underlying model.

The function outputs the resulting average log likelihood over all folds for
each parameter as well as the value of the parameter with which the maximum
likelihood was achieved.
}
