% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lslm_boost.R
\name{lslm_boost}
\alias{lslm_boost}
\title{Componentwise gradient boosting in location-scale regression}
\usage{
lslm_boost(
  location,
  scale = ~1,
  data = environment(location),
  light = FALSE,
  mstop = 1000,
  nu_b = 0.1,
  nu_g = 0.01
)
}
\arguments{
\item{location}{A two-sided formula with the response variable on the LHS
and the predictor for the mean on the RHS.}

\item{scale}{A one-sided formula with the predictor for the standard
deviation on the RHS.}

\item{data}{A data frame (or list or environment) in which to evaluate
the \code{location} and \code{scale} formulas.}

\item{light}{If \code{TRUE}, the design matrices are removed from the estimated
model to save some memory. Has to be \code{FALSE} if the model is
to be used in cross validation.}

\item{mstop}{Integer denoting the number of iterations the algorithm. Defaults to 1000.}

\item{nu_b}{Learning rate for the location coefficients. Defaults to 0.1.}

\item{nu_g}{Learning rate for the scale coefficients. Defaults to 0.01.}
}
\description{
The location-scale regression model assumes a normally distributed response
variable with one linear predictor for the mean (= the location) and one for
the standard deviation (= the scale). The standard deviation is mapped to
the linear predictor through a log link.

This function sets up the model object and estimates it with componentwise gradient boosting.
}
