% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulating data for Gaussian Location-Scale Regression models}
\usage{
simulate_data(n_obs, n_var, seed = NULL, frac_beta = 0.25, frac_gamma = 0.25)
}
\arguments{
\item{n_obs}{number of observations to be generated.}

\item{n_var}{number of covariates for the linear models.}

\item{seed}{seed for random number generator.}

\item{frac_beta}{fraction of location coefficients that are to be set to 0.}

\item{frac_gamma}{fraction of scale coefficients that are to be set to 0.}
}
\value{
A list containing the objects:
\itemize{
\item y - gaussian random variable with mean X \%\emph{\% beta_coef and standard deviation
exp(X \%}\% gamma_coef)
\item X - matrix of covariates used to generate y
\item beta_coef - coefficients for location parameter
\item gamma_coef - coefficients for scale parameter
}
}
\description{
Simulating a normally distributed random variable with a linear model as mean and a
log-linear model as standard deviation.
}
